filter/filter_test.go
package filter

import (
	"net/http"
	"testing"

	"github.com/google/martian/header"
	"github.com/google/martian/martiantest"
)

func TestRequestCondition(t *testing.T) {
	hm := header.NewMatcher("Martian-Testing", "true")
	tm := martiantest.NewModifier()

	f := Filter{}
	f.SetRequestCondition(hm, tm)

	tt := []struct {
		name   string
		values []string
		want   bool
	}{
		{
			name:   "Martian-Production",
			values: []string{"true"},
			want:   false,
		},
		{
			name:   "Martian-Testing",
			values: []string{"see-next-value", "true"},
			want:   true,
		},
	}

	for i, tc := range tt {
		req, err := http.NewRequest("GET", "/", nil)
		if err != nil {
			t.Fatalf("http.NewRequest(): got %v, want no error", err)
		}

		req.Header[tc.name] = tc.values

		if err := f.ModifyRequest(req); err != nil {
			t.Fatalf("%d. ModifyRequest(): got %v, want no error", i, err)
		}

		if tm.RequestModified() != tc.want {
			t.Errorf("%d. tm.RequestModified(): got %t, want %t", i, tm.RequestModified(), tc.want)
		}
	}
}
